Using keysize and extracted legacy code into its own file.
