+++ security/vault.go
@@ -5,7 +5,6 @@ import (
 	"crypto/aes"
 	"crypto/cipher"
 	"crypto/sha256"
-	"encoding/base64"
 	"encoding/binary"
 	"encoding/hex"
 	"errors"
@@ -20,10 +19,8 @@ import (
 )
 
 const (
-	vaultName        = ".gaia_vault"
-	keySize          = 32
-	secretCheckKey   = "GAIA_CHECK_SECRET"
-	secretCheckValue = "!CHECK_ME!"
+	vaultName = ".gaia_vault"
+	keySize   = 32
 )
 
 // VaultAPI defines a set of apis that a Vault must provide in order to be a Gaia Vault.
@@ -87,7 +84,7 @@ func NewVault(ca CAAPI, storer VaultStorer) (*Vault, error) {
 	if err != nil {
 		return nil, err
 	}
-	if len(data) < 32 {
+	if len(data) < keySize {
 		return nil, errors.New("key lenght should be longer than 32")
 	}
 	h := sha256.New()
@@ -270,44 +267,6 @@ func (v *Vault) decrypt(encodedData []byte) ([]byte, error) {
 	return plaintext, nil
 }
 
-func (v *Vault) legacyDecrypt(data []byte) ([]byte, error) {
-	if len(data) < 1 {
-		gaia.Cfg.Logger.Info("the vault is empty")
-		return []byte{}, nil
-	}
-	paddedPassword := v.pad(v.cert)
-	ci := base64.URLEncoding.EncodeToString(paddedPassword)
-	block, err := aes.NewCipher([]byte(ci[:aes.BlockSize]))
-	if err != nil {
-		return []byte{}, err
-	}
-
-	decodedMsg, err := base64.URLEncoding.DecodeString(string(data))
-	if err != nil {
-		return []byte{}, err
-	}
-
-	if (len(decodedMsg) % aes.BlockSize) != 0 {
-		return []byte{}, errors.New("blocksize must be multiple of decoded message length")
-	}
-
-	iv := decodedMsg[:aes.BlockSize]
-	msg := decodedMsg[aes.BlockSize:]
-
-	cfb := cipher.NewCFBDecrypter(block, iv)
-	cfb.XORKeyStream(msg, msg)
-
-	unpadMsg, err := v.unpad(msg)
-	if err != nil {
-		return []byte{}, err
-	}
-
-	if !bytes.Contains(unpadMsg, []byte(secretCheckValue)) {
-		return []byte{}, errors.New("possible mistyped password")
-	}
-	return unpadMsg, nil
-}
-
 // ParseToMap will update the Vault data map with values from
 // an encrypted file content.
 func (v *Vault) parseToMap(data []byte) error {
@@ -335,23 +294,4 @@ func (v *Vault) parseFromMap() []byte {
 	}
 
 	return bytes.Join(data, []byte("\n"))
-}
-
-// Pad pads the src with 0x04 until block length.
-func (v *Vault) pad(src []byte) []byte {
-	padding := aes.BlockSize - len(src)%aes.BlockSize
-	padtext := bytes.Repeat([]byte{byte(padding)}, padding)
-	return append(src, padtext...)
-}
-
-// Unpad removes the padding from pad.
-func (v *Vault) unpad(src []byte) ([]byte, error) {
-	length := len(src)
-	unpadding := int(src[length-1])
-
-	if unpadding > length {
-		return nil, errors.New("possible mistyped password")
-	}
-
-	return src[:(length - unpadding)], nil
 }
